pyproject.toml
[project]
name = "eoairflow"
version = "0.1.0"
description = "Airflow workflows defined using JSON graphs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow>=2.8.0",
    "apache-airflow-providers-docker>=3.0.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
